requirements.txt
aiogram
asyncpg
python-dotenv
stripe
flask
